First design of new mid-level interface style
